Move/refactory devices/uuids methods
